feat: ability to disable keymaps

+++ lua/recorder.lua
@@ -353,12 +353,22 @@ function M.setup(userConfig)
 	local icon = config.useNerdfontIcons and " " or ""
 	local dapSharedIcon = config.useNerdfontIcons and " /  " or ""
 
-	keymap("n", toggleKey, toggleRecording, { desc = icon .. "Start/Stop Recording" })
-	keymap("n", config.mapping.switchSlot, switchMacroSlot, { desc = icon .. "Switch Macro Slot" })
-	keymap("n", config.mapping.editMacro, editMacro, { desc = icon .. "Edit Macro" })
-	keymap("n", config.mapping.yankMacro, yankMacro, { desc = icon .. "Yank Macro" })
-	-- stylua: ignore
-	keymap("n", config.mapping.deleteAllMacros, deleteAllMacros, { desc = icon .. "Delete All Macros" })
+	if toggleKey then
+		keymap("n", toggleKey, toggleRecording, { desc = icon .. "Start/Stop Recording" })
+	end
+	if config.mapping.switchSlot then
+		keymap("n", config.mapping.switchSlot, switchMacroSlot, { desc = icon .. "Switch Macro Slot" })
+	end
+	if config.mapping.editMacro then
+		keymap("n", config.mapping.editMacro, editMacro, { desc = icon .. "Edit Macro" })
+	end
+	if config.mapping.yankMacro then
+		keymap("n", config.mapping.yankMacro, yankMacro, { desc = icon .. "Yank Macro" })
+	end
+	if config.mapping.deleteAllMacros then
+		-- stylua: ignore
+		keymap("n", config.mapping.deleteAllMacros, deleteAllMacros, { desc = icon .. "Delete All Macros" })
+	end
 
 	-- (experimental) if true, nvim-recorder and dap will use shared keymaps:
 	-- 1) `addBreakPoint` will map to `dap.toggle_breakpoint()` outside
@@ -367,11 +377,16 @@ function M.setup(userConfig)
 	-- dap-breakpoint. If there is no dap breakpoint, will play the current
 	-- macro-slot instead
 	dapSharedKeymaps = config.dapSharedKeymaps or false
-	local breakPointDesc = dapSharedKeymaps and dapSharedIcon .. "Breakpoint"
-		or icon .. "Insert Macro Breakpoint."
-	keymap("n", breakPointKey, addBreakPoint, { desc = breakPointDesc })
-	local playDesc = dapSharedKeymaps and dapSharedIcon .. "Continue/Play" or icon .. "Play Macro"
-	keymap("n", config.mapping.playMacro, playRecording, { desc = playDesc })
+	if config.mapping.addBreakPoint then
+		local breakPointDesc = dapSharedKeymaps and dapSharedIcon .. "Breakpoint"
+			or icon .. "Insert Macro Breakpoint."
+		keymap("n", breakPointKey, addBreakPoint, { desc = breakPointDesc })
+	end
+
+	if config.mapping.playMacro then
+		local playDesc = dapSharedKeymaps and dapSharedIcon .. "Continue/Play" or icon .. "Play Macro"
+		keymap("n", config.mapping.playMacro, playRecording, { desc = playDesc })
+	end
 end
 
 --------------------------------------------------------------------------------
